Moved all source files into src. 1.1.0 Windows Pre-release
Moved most files to  /src
Fixed up imports to declared from src.
Fixed updater versioning check
