+++ StellarisChecksumPatcher/main.py
@@ -1,5 +1,5 @@
 # 3rd-party
-from UI import stellaris_checksum_patcher_gui
+from src.UI import stellaris_checksum_patcher_gui
 
 if __name__ == '__main__':
     w = stellaris_checksum_patcher_gui.StellarisChecksumPatcherGUI()
