+++ StellarisChecksumPatcher/src/UI/StellarisChecksumPatcherUI.ui
