+++ StellarisChecksumPatcher/src/UI/ui_auto_builder.py
