+++ StellarisChecksumPatcher/src/UI/ui_icons/save_patch_icon.png
