+++ StellarisChecksumPatcher/src/UI/ui_utils.py
