+++ StellarisChecksumPatcher/src/achievements/achievements.txt
