+++ StellarisChecksumPatcher/src/patchers/HexPatcher-deprecated.py
